refactor, consolidate terminal command mirrors
