+++ src/vs/workbench/contrib/terminal/browser/chatTerminalCommandMirror.ts
@@ -15,6 +15,57 @@ import { PANEL_BACKGROUND } from '../../../common/theme.js';
 import { IContextKeyService } from '../../../../platform/contextkey/common/contextkey.js';
 import { ChatContextKeys } from '../../chat/common/chatContextKeys.js';
 import { editorBackground } from '../../../../platform/theme/common/colorRegistry.js';
+import { Color } from '../../../../base/common/color.js';
+import type { IChatTerminalToolInvocationData } from '../../chat/common/chatService.js';
+import type { IColorTheme } from '../../../../platform/theme/common/themeService.js';
+
+function getChatTerminalBackgroundColor(theme: IColorTheme, contextKeyService: IContextKeyService, storedBackground?: string): Color | undefined {
+	if (storedBackground) {
+		const color = Color.fromHex(storedBackground);
+		if (color) {
+			return color;
+		}
+	}
+
+	const terminalBackground = theme.getColor(TERMINAL_BACKGROUND_COLOR);
+	if (terminalBackground) {
+		return terminalBackground;
+	}
+
+	const isInEditor = ChatContextKeys.inChatEditor.getValue(contextKeyService);
+	return theme.getColor(isInEditor ? editorBackground : PANEL_BACKGROUND);
+}
+
+/**
+ * Base class for detached terminal mirrors.
+ * Handles attaching to containers and managing the detached terminal instance.
+ */
+abstract class DetachedTerminalMirror extends Disposable {
+	private _detachedTerminal: Promise<IDetachedTerminalInstance> | undefined;
+	private _attachedContainer: HTMLElement | undefined;
+
+	protected _setDetachedTerminal(detachedTerminal: Promise<IDetachedTerminalInstance>): void {
+		this._detachedTerminal = detachedTerminal.then(terminal => this._register(terminal));
+	}
+
+	protected async _getTerminal(): Promise<IDetachedTerminalInstance> {
+		if (!this._detachedTerminal) {
+			throw new Error('Detached terminal not initialized');
+		}
+		return this._detachedTerminal;
+	}
+
+	protected async _attachToContainer(container: HTMLElement): Promise<IDetachedTerminalInstance> {
+		const terminal = await this._getTerminal();
+		container.classList.add('chat-terminal-output-terminal');
+		const needsAttach = this._attachedContainer !== container || container.firstChild === null;
+		if (needsAttach) {
+			terminal.attachToElement(container);
+			this._attachedContainer = container;
+		}
+		return terminal;
+	}
+}
 
 export async function getCommandOutputSnapshot(
 	xtermTerminal: XtermTerminal,
@@ -91,10 +142,7 @@ interface IDetachedTerminalCommandMirror {
  * Mirrors a terminal command's output into a detached terminal instance.
  * Used in the chat terminal tool progress part to show command output for example.
  */
-export class DetachedTerminalCommandMirror extends Disposable implements IDetachedTerminalCommandMirror {
-	private _detachedTerminal: Promise<IDetachedTerminalInstance>;
-	private _attachedContainer?: HTMLElement;
-
+export class DetachedTerminalCommandMirror extends DetachedTerminalMirror implements IDetachedTerminalCommandMirror {
 	constructor(
 		private readonly _xtermTerminal: XtermTerminal,
 		private readonly _command: ITerminalCommand,
@@ -102,17 +150,23 @@ export class DetachedTerminalCommandMirror extends Disposable implements IDetach
 		@IContextKeyService private readonly _contextKeyService: IContextKeyService,
 	) {
 		super();
-		this._detachedTerminal = this._createTerminal();
+		const processInfo = this._register(new DetachedProcessInfo({ initialCwd: '' }));
+		const capabilities = this._register(new TerminalCapabilityStore());
+		this._setDetachedTerminal(this._terminalService.createDetachedTerminal({
+			cols: this._xtermTerminal.raw!.cols,
+			rows: 10,
+			readonly: true,
+			processInfo,
+			disableOverviewRuler: true,
+			capabilities,
+			colorProvider: {
+				getBackgroundColor: theme => getChatTerminalBackgroundColor(theme, this._contextKeyService),
+			},
+		}));
 	}
 
 	async attach(container: HTMLElement): Promise<void> {
-		const terminal = await this._detachedTerminal;
-		container.classList.add('chat-terminal-output-terminal');
-		const needsAttach = this._attachedContainer !== container || container.firstChild === null;
-		if (needsAttach) {
-			terminal.attachToElement(container);
-			this._attachedContainer = container;
-		}
+		await this._attachToContainer(container);
 	}
 
 	async renderCommand(): Promise<{ lineCount?: number } | undefined> {
@@ -123,7 +177,7 @@ export class DetachedTerminalCommandMirror extends Disposable implements IDetach
 		if (!vt.text) {
 			return { lineCount: 0 };
 		}
-		const detached = await this._detachedTerminal;
+		const detached = await this._getTerminal();
 		detached.xterm.clearBuffer();
 		detached.xterm.clearSearchDecorations?.();
 		await new Promise<void>(resolve => {
@@ -131,30 +185,102 @@ export class DetachedTerminalCommandMirror extends Disposable implements IDetach
 		});
 		return { lineCount: vt.lineCount };
 	}
+}
 
-	private async _createTerminal(): Promise<IDetachedTerminalInstance> {
+/**
+ * Mirrors a terminal output snapshot into a detached terminal instance.
+ * Used when the terminal has been disposed of but we still want to show the output.
+ */
+export class DetachedTerminalSnapshotMirror extends DetachedTerminalMirror {
+	private _output: IChatTerminalToolInvocationData['terminalCommandOutput'] | undefined;
+	private _container: HTMLElement | undefined;
+	private _dirty = true;
+	private _lastRenderedLineCount: number | undefined;
+
+	constructor(
+		output: IChatTerminalToolInvocationData['terminalCommandOutput'] | undefined,
+		private readonly _getTheme: () => IChatTerminalToolInvocationData['terminalTheme'] | undefined,
+		@ITerminalService private readonly _terminalService: ITerminalService,
+		@IContextKeyService private readonly _contextKeyService: IContextKeyService,
+	) {
+		super();
+		this._output = output;
 		const processInfo = this._register(new DetachedProcessInfo({ initialCwd: '' }));
-		const capabilities = this._register(new TerminalCapabilityStore());
-		const detached = await this._terminalService.createDetachedTerminal({
-			cols: this._xtermTerminal.raw!.cols,
+		this._setDetachedTerminal(this._terminalService.createDetachedTerminal({
+			cols: 80,
 			rows: 10,
 			readonly: true,
 			processInfo,
 			disableOverviewRuler: true,
-			capabilities,
 			colorProvider: {
 				getBackgroundColor: theme => {
-					const terminalBackground = theme.getColor(TERMINAL_BACKGROUND_COLOR);
-					if (terminalBackground) {
-						return terminalBackground;
-					}
-					// Use editor background when in chat editor, panel background otherwise
-					const isInEditor = ChatContextKeys.inChatEditor.getValue(this._contextKeyService);
-					return theme.getColor(isInEditor ? editorBackground : PANEL_BACKGROUND);
-				},
+					const storedBackground = this._getTheme()?.background;
+					return getChatTerminalBackgroundColor(theme, this._contextKeyService, storedBackground);
+				}
 			}
-		});
-		return this._register(detached);
+		}));
 	}
 
+	public setOutput(output: IChatTerminalToolInvocationData['terminalCommandOutput'] | undefined): void {
+		this._output = output;
+		this._dirty = true;
+	}
+
+	public async attach(container: HTMLElement): Promise<void> {
+		await this._attachToContainer(container);
+		this._container = container;
+		this._applyTheme(container);
+	}
+
+	public async render(): Promise<{ lineCount?: number } | undefined> {
+		const output = this._output;
+		if (!output) {
+			return undefined;
+		}
+		if (!this._dirty) {
+			return { lineCount: this._lastRenderedLineCount ?? output.lineCount };
+		}
+		const terminal = await this._getTerminal();
+		terminal.xterm.clearBuffer();
+		terminal.xterm.clearSearchDecorations?.();
+		if (this._container) {
+			this._applyTheme(this._container);
+		}
+		const text = output.text ?? '';
+		const lineCount = output.lineCount ?? this._estimateLineCount(text);
+		if (!text) {
+			this._dirty = false;
+			this._lastRenderedLineCount = lineCount;
+			return { lineCount: 0 };
+		}
+		await new Promise<void>(resolve => terminal.xterm.write(text, resolve));
+		this._dirty = false;
+		this._lastRenderedLineCount = lineCount;
+		return { lineCount };
+	}
+
+	private _estimateLineCount(text: string): number {
+		if (!text) {
+			return 0;
+		}
+		const sanitized = text.replace(/\r/g, '');
+		const segments = sanitized.split('\n');
+		const count = sanitized.endsWith('\n') ? segments.length - 1 : segments.length;
+		return Math.max(count, 1);
+	}
+
+	private _applyTheme(container: HTMLElement): void {
+		const theme = this._getTheme();
+		if (!theme) {
+			container.style.removeProperty('background-color');
+			container.style.removeProperty('color');
+			return;
+		}
+		if (theme.background) {
+			container.style.backgroundColor = theme.background;
+		}
+		if (theme.foreground) {
+			container.style.color = theme.foreground;
+		}
+	}
 }
